feat: add validation to make sure user cannot delete others links

+++ src/services/links/delete-link.service.ts
@@ -1,15 +1,22 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { LinksRepository } from '../../database/contracts/contract-links-repository';
 
 type Request = {
   id: string;
+  userId?: string;
 };
 
 @Injectable()
 export class DeleteLinkService {
   constructor(private linksRepository: LinksRepository) {}
 
-  async execute({ id }: Request): Promise<void> {
+  async execute({ id, userId }: Request): Promise<void> {
+    const link = await this.linksRepository.findById(id);
+    if (link && link.userId !== userId) {
+      throw new UnauthorizedException(
+        'You are not allowed to delete this link.',
+      );
+    }
     await this.linksRepository.delete(id);
   }
 }
